[ios] Cleanup //ios/chrome/app/application_delegate dependencies
With https://crrev.com/c/6011911 AppInitStage and ProfileInitStage
are decoupled, and most of the logic in AppStage has moved instead
to MailController and ProfileController.

This means that many of the dependencies are obsolete and can now
be removed, and the target application_delegate_internal can be
split which allow to remove the circular dependencies. This makes
the app_state_header target redundant since it is possible to
depend directly on //ios/chrome/app/applicaton_delegate:app_state
now.

Bug: 41492447
Change-Id: I5529bdd6a4b6835ab9ee0f9f5cd67fe48ec638fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6026033
Reviewed-by: Aliona Dangla <alionadangla@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1384256}