[@scope] Bucket :scope as <scope-start> when possible
This CL implements the same bucketing optimization for :scope
as we do  for :is()/:where(): whenever we have a <scope-start>
with exactly one selector, we bucket :scope as that selector instead.

This optimization arguably becomes slightly more relevant with
CL:6020075, since that CL makes it more ergonomic to apply
declarations directly to the scoping root (effectively matching
like :where(:scope)).

Rename HasOneSelector to IsSingleComplexSelector, to make it more clear
that we're interested in the number of complex selectors (vs. simple
selectors). This is especially relevant for the new static version
of this function.

Bug: 351045927
Change-Id: I7c25358e4a448cda15c89d44422b558f1900ae02
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6028416
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1384301}