[iOS][LO][Clobber][3/N] Add LensSubNavigationItem
Currently, LensOverlayNavigationManager stores the URL for each sub
navigation. (a sub navigation is a navigation with the same lens overlay
result).

This CL adds LensSubNavigationItem, to store the omnibox text with the
URL for each sub navigation. This allows the omnibox text to be restored
when going to a previous navigation.

Bug: 375194454
Change-Id: I8b797f05d426e37dfd1597a394675b6184445013
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6017852
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Auto-Submit: Christian Xu <christianxu@chromium.org>
Commit-Queue: Christian Xu <christianxu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1384340}