[ios] Stop overriding policy pref service in VariationsService
By default VariationsService looks in the local state PrefService,
but there is a setter to override the PrefService used to look up
the preference kVariationsRestrictParameter.

This is only possible to do if there is a single profile, which is
no longer the case when there are multiple profiles. Since the pref
is define both in the local state and profile PrefServices, use the
value from local state.

Additionally it look like this preference is never written, so it
should not cause any change.

Fixed: 343152173
Change-Id: I67f61dbaf10c49771d099535c40b3e59913713fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6030909
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1384375}