[NavCapture]: Re-enable tests marked as flaky.
We've recently made an effort to stamp out flakiness in our testing
infrastructure. We therefore have good reason to believe that these tests now pass. This is an effort to put that to the test.

Sheriffs: If one or two of those tests starts showing signs of
flakiness, please add the relevant string back (don't revert the
whole CL).

Bug: 377425233
Change-Id: Ic545a8d443b8a8fd292619c6b02082a8a9388008
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6030429
Reviewed-by: Dibyajyoti Pal <dibyapal@chromium.org>
Commit-Queue: Finnur Thorarinsson <finnur@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1384379}