Allow collect_warnings.py to return codesearch links if requested
While triaging warnings, I frequently found myself wishing I had a
direct way to go from the name of the file in the output of
collect_warnings.py to viewing it in codesearch. This CL allows that
by adding a new option to write codesearch links directly in the output. With a suitable editor, these links can be followed with a single click.

Since the links make the lines longer, this option also changes the
output format slightly when in summarize mode to make it more readable. A possible direction for the future would be to output
markdown and let it handle the links, but that seems like overkill
for the moment.

Change-Id: I707b12597bd5a27f4118006f7b90198e3985fb07
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6013102
Commit-Queue: Devon Loehr <dloehr@google.com>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1384382}