Detect operator[] on empty fixed-size spans at compile time.
This already happened with front() and back().

This requires changing the implementation of operator<=>() to something
more like the description in the standard, so it doesn't attempt to call
operator[]() on a zero-extent span.

Bug: 364987728
Change-Id: I32fbcd4d46656d4c95fd9a5513e32a39f0b26678
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6024195
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1384459}