[lensoverlay] Add UMA for document size
Adds a new UMA to track the size of the document where the contextual search box was shown. The size is sliced by content type (PDF, HTML, or plain text). For web, both the size of the innerText and the innerHtml are logged to help estimate the delta between the two.

This change also fixes a bug where we sometimes update the page content
multiple times for a single page that didn't change.

Bug: 366028802
Change-Id: I7c7b799ed798a2eb2c6bd67bb80c0a38eed2cdc8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6022581
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Reviewed-by: Ali Stanfield <stanfield@google.com>
Commit-Queue: Duncan Mercer <mercerd@google.com>
Cr-Commit-Position: refs/heads/main@{#1384472}