Mac immersive FS: keep thin controller in window permanently
To properly set up autohiding, macOS does a bunch of bookkeeping
during the fullscreen transition. One of the things it checks is if
the window has any titlebar accessory view controllers; if not, it
decides autohiding is not necessary and caches a bunch of stuff. One
of the things that gets cached is the reported reveal amount, which
gets stuck at 100%.

In PWAs this creates a race condition: we round trip to the browser
before we enable immersive fullscreen. This means that by the time
we add our titlebar view controllers, the transition is finished.
Autohiding still *works* (presumably since it does its own check),
but the state that we look at for positioning the overlay window is
stuck (making the area it covers uninteractable).

As it happens, we have an inert titlebar view controller to work
around a different bug. This change adds a feature to move ownership
of this controller to the browser window and keep it there permanently.
This fixes the check and all is well (or is it? That's why it's a
feature!)

Low-Coverage-Reason: OTHER Feature disabled by default.
Bug: 373722654
Change-Id: I3893343dee444cb9512cf7a3e5cf8d3ff2b53092
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6023424
Reviewed-by: Keren Zhu <kerenzhu@chromium.org>
Commit-Queue: Leonard Grey <lgrey@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1384483}