Make TriggerManager::GetSBErrorDisplayOptions() return smaller struct
Currently GetSBErrorDisplayOptions() returns a large struct of type
security_interstitials::BaseSafeBrowsingErrorUI::SBErrorDisplayOptions.
However, the calls of that method only end up using 3 of the struct
fields.

This CL replaces that function with GetDataCollectionPermissions(),
which returns a smaller struct DataCollectionPermissions that only
contains the 3 relevant fields. TriggerManager methods that used the
struct SBErrorDisplayOptions now use DataCollectionPermissions instead.

The two blocking page classes used a separate BaseSafeBrowsingErrorUI
helper method to thread through a SBErrorDisplayOptions to methods on
the TriggerManager. These helper method calls are left as-is, and the
returned struct is converted into a DataCollectionPermissions after.

Fixed: 345125741
Change-Id: Icf93c4c344fb774b66d93f9e8507fc46772c8387
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6026188
Reviewed-by: Zack Han <zackhan@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Commit-Queue: thefrog <thefrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1384499}