Create Glic Window Manager and Controller.
The Window Manager ensures that only one instance of the Glic panel is present at a time. It will also handle which profile to choose and to show the Profile Picker when needed.

The Window Controller creates and owns the Glic Widget, and will also be owned by a future Glic Keyed-Service. Currently empty but soon-to-be implemented.

Change-Id: I771e42b1c58f7fe3e3e42f640193f5330cff8ad9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6021733
Reviewed-by: Anthony Cui <cuianthony@chromium.org>
Reviewed-by: Ian Wells <iwells@chromium.org>
Commit-Queue: Sana Akbani <sanaakbani@google.com>
Cr-Commit-Position: refs/heads/main@{#1384506}