Fix bug that allowed actual quota to still be leaked even with static quota enabled.
For buckets with a requested quota greater than the StorageKey quota,
quota gets capped at the StorageKey quota. This meant that even
with the kStaticStorageQuota flag enabled, if you
created a bucket with a high requested quota you'd still be able to
detect the user's browsing mode based on what it gets capped to.

This changes the kStaticStorageQuota implementation so that whatever quota value the bucket creator requested gets reported by estimate() even if in reality it would get capped.

Bug: 369865059
Change-Id: Ic55da0ba9c36850c50abc46e0ece6334e8cfd2a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6020091
Commit-Queue: Reema A <reemaa@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1384536}