webview: record use from non-app UIDs in UMA.
When WebView is used from a non-app (privileged/system) UID, record this
in a UMA histogram so we can investigate where this is happening and
why. This is only implemented for SDK30+ as the relevant Android API was
not available before then, and we aren't really interested in old
Android versions.

Change-Id: I1ca74ed368e80176ba36419e2c3abcc8217115c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6017447
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Auto-Submit: Richard (Torne) Coles <torne@chromium.org>
Commit-Queue: Richard (Torne) Coles <torne@chromium.org>
Reviewed-by: Caitlin Fischer <caitlinfischer@google.com>
Cr-Commit-Position: refs/heads/main@{#1384544}