Begin using SeedReaderWriter to get seeds.
This CL (CL 2) is part of the following series of changes to add support
for reading variations seeds from the new seed file.

CL 1. Implements (but does not yet use) the ability to read seeds from
the seed file: crrev.com/c/5955287

CL 2. Begins using a SeedReaderWriter to read seed.

This CL also begins using safe seed sentinel values in the seed file.
This was previously not possible as there was no way to get
non-encoded seed data stored in the seed file when the logic calls for
a comparison of the latest seed to the safe seed in sentinel-related
code.

Bug: 374947675
Change-Id: If76be3a4a084dd5ce5bb966037cce8281ae3adf4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6004700
Reviewed-by: Caitlin Fischer <caitlinfischer@google.com>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Cristian Lopez <cristiandlopez@google.com>
Cr-Commit-Position: refs/heads/main@{#1384564}