firmware_update: hide restart buttons on failure dialog
The recently added restart buttons are only visible when `isUpdateDone`
and `updateRequiresRestart` is true. However, `isUpdateDone` is also
true when a firmware update fails, thus the restart buttons erroneously
appear on the failure dialog. This changes fixes that and adds the test
that would've caught this in CQ.

Test: out/Default/browser_tests --gtest_filter=FirmwareUpdateAppBrowserTest.FirmwareUpdateDialog
Bug: b:379342936
Change-Id: I4538ce9d70178741c78b72d74b0db7cf1224d261
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6027313
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Reviewed-by: Michael Checo <michaelcheco@google.com>
Commit-Queue: Andre Braga <andrebraga@google.com>
Cr-Commit-Position: refs/heads/main@{#1384587}