[TG Sync] Make SavedTabGroupWebContentsListener a TabFeature
This CL makes SavedTabGroupWebContentsListener a TabFeature.
With this, the ownership of STGWCL moves to the TabModel itself
instead of manually managing it in LocalTabGroupListener.
This simplifies the design and aligns better with implementation
in Android and iOS.
Here are some of the details.

SavedTabGroupWebContentsListener:
1. STGWCL is 1:1 with a tabs::TabModel. Created for every
normal tab in the regular profile.
2. DidFinishNavigation filters out the navigations for tabs that
aren't in tab groups as early return, hence it's equivalent
to the existing design.
3. We improve the saved_group() method in STGWCL to get the
SavedTabGroup corresponding to the local tab group ID, hence
it's better than iterating over all groups.

LocalTabGroupListener:
1. It doesn't maintain a lister map of STGWCL any more. STGWCL is
directly obtained from the tab feature of the TabModel which is
normally an argument to its public methods. This aligns with
iOS/Android implementation.
2. Earlier, in every public method we were checking whether
the incoming event corresponds to a STGWCL in the listener map.
This is replaced by checking if the LocalTabID (tab handle) of
the TabModel exists in the SavedTabGroup or not. This is
functionally equivalent.

Bug: 359982911
Change-Id: Ifc4d071c6b7f84ba485e5f6ccf967654dab7e61f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5988894
Reviewed-by: Darryl James <dljames@chromium.org>
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Reviewed-by: David Pennington <dpenning@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1384590}