coral: Update desk preview view at the end of expanding desks bar
When creating a new desk for coral group from the zero state desks bar,
at the end of new desk creation, the desk previews for both desks show
empty content. The reason is when performing the animation of expanding
desks bar, we manually disable the layout of desks bar to achieve some
performance improvement and enable the layout at the end of the
animation.

Code links:
1. Disable layout:
https://source.chromium.org/chromium/chromium/src/+/main:ash/wm/desks/desk_mini_view_animations.cc;drc=0ecb488b863be0b6c031e44eafcca89b39619522;l=220
2. Enable layout at the end of the animation:
https://source.chromium.org/chromium/chromium/src/+/main:ash/wm/desks/desk_mini_view_animations.cc;drc=0ecb488b863be0b6c031e44eafcca89b39619522;l=226

Hence we manually call `InvalidateLayout()` on desks bar to ensure that
all its child views are painted correctly at the end of the animation.

Screen records:
1. Before fix:
https://screencast.googleplex.com/cast/NjA0MzU4Mjg2MjQ1ODg4MHxhZGYxMTBiYy1mNQ
2. After fix:
https://screencast.googleplex.com/cast/NjAwNDE5ODY4MzQ0MzIwMHw4OGU1YzExYS1kNQ

Bug=b:378558824

Change-Id: I926e065352fb3b8d19463b434882ccb148a38b49
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6032010
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Eric Sum <esum@google.com>
Commit-Queue: Connie Xu <conniekxu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1384611}