Set default-dir, suggested-name in EXTRAS_INITIAL_URI, EXTRAS_TITLE
File System Access allows file choosers to set the default_directory
to one of the chosen directory file handles. Android supports this
via EXTRAS_INITIAL_URI.

JS API showSaveFilePicker() takes an optional suggestedName for the new
file which should be added into EXTRAS_TITLE in CREATE_DOCUMENT intent
to auto-fill a suggested name in save-as.

FileSystemChooser::Options::default_path() is created as
default_directory.Append(suggested_name) where both can be empty.
It is updated to always end with a path separator when suggested_name
is empty so that it can be reliably parsed back into default_directory
and suggested_name after default_path is passed as a param to
SelectFileDialgo::SelectFile(). The alternative of modifying the
existing SelectFile() default_path into 2 fields would not work since
many callsites have only the default_path which is ambiguous as to
whether it is a directory, or suggested-name or combination.

Bug: 379140421
Change-Id: Ife6ce61471c95acd4ef763ed6e570e063afc114c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6024565
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Joel Hockey <joelhockey@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Nathan Memmott <memmott@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1384627}