Move dlp_policy_event.proto to synced directory
This will become the source of truth for the copy of this proto in
serverside code and be synced daily. A few changes to the proto, moving
from proto2 -> proto3 since that matches the serverside proto more
closely, and the only use case in Chromium code is unaffected by the
proto2 -> proto3 change.

Change-Id: I3315a48785d49c8ad3aa0d48c50658dfd2f24f54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6015313
Commit-Queue: Shanthanu Bhardwaj <xanth@google.com>
Reviewed-by: Dominique Fauteux-Chapleau <domfc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1384644}