Fix system bars coloring bug by delaying creation of InsetObserver
It looks like initializing the InsetObserver too early in ChromeBaseAppCompatActivity#onCreate causes some coloring issues with the system bars. This change delays the initialization of the InsetObserver until the root issue can be identified and fixed, which incidentally also forces the EdgeToEdgeStateProvider, which is needed by the ModalDialogManager, to be created outside of the EdgeToEdgeManager.

Ideally the root issue will be fixed in future CLs and the EdgeToEdgeStateProvider can be returned to the EdgeToEdgeManager.

Bug: 379153231
Change-Id: I6051d8c0411b5a593326ab7fdec05939e255171a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6026294
Commit-Queue: Charles Hager <clhager@google.com>
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1384660}