metrics: Adding CrOSPreConsentManager
The CrOSPreConsentManager allows for metrics to be enabled in on
ChromeOS devices, primarily during OOBE when the user hasn't consented.
The object must be enabled by externally first; on subsequence restarts
it will be enabled by default according to the preference. Additionally,
this object will not be created if the user has consented to metrics
(negatively or positively) determined by a file created by this object
disabled. This object is disabled when consent has been given by user.

Introduces a new feature ash::features::kOobePreConsentMetrics for
this functionality. This functionality is unused and will be integrated
in a follow-up CL.

Change-Id: I7cbe2ce94d4fa6740a0b991344c667a2b728f267
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5949603
Commit-Queue: Andrew Bregger <andrewbregger@google.com>
Reviewed-by: Jong Ahn <jongahn@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1384695}