[AppIntegration] Don't set schema repeatedly once it has been set.
After creating a database for tab donation, Chrome only need to set
the schema which is the format of the data to send once. In this CL,
we save the result of setting schema into the SharedPreference to
prevent setting the schema on every launch.

Bug: 376549664
Change-Id: I666d65457aa24077af68f5f92101dee2c4898cc5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6031510
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Xi Han <hanxi@chromium.org>
Reviewed-by: Gang Wu <gangwu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1384703}