coral: Save apps from group into coral desk
Uses the existing restore data collector to save the apps from the
active desk. Passes a map as an allow list so we only save the apps in
the group. The restore data collector is built to be async (although
with lacros gone, in theory it should be sync), so we need to shuffle
some existing logic to follow this pattern.

Test: manual
Test: manually checked disk contents
Test: ash_unittests CoralSavedGroupTest.SaveAppsInGroup
Bug: 365839564
Change-Id: I8bf0aa220291c037131a8550b014a1fd7d10b449
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6032298
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1384766}