iwa: version pinning for update discovery
Allow administrators to pin an already installed, managed IWA to a
specific version. During the search for possible updates, the pinned_version field from the IsolatedWebAppInstallForceList
policy will be checked for existence.
The update will only be scheduled when the pinned_version is higher
than the current one and is available in the current update channel.
By default, no downgrading is possible.

Bug: b:321672070
Change-Id: Ib9a02635a5f8a09ab0f6430a1ceeb78d2d84787d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6003153
Commit-Queue: Magdalena Skarbińska <mskarbinska@google.com>
Reviewed-by: Andrew Rayskiy <greengrape@google.com>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Simon Hangl <simonha@google.com>
Cr-Commit-Position: refs/heads/main@{#1384823}