[Tab Groups] Add new observer method for SavedTabGroup migration
The new method is called for the transition from saved to shared tab
group (both remote and local). This results in a deletion of previous
saved tab group from the service, creation of a new shared group, and
migrating all the local IDs (both tab and group).

Note that OnTabGroupUpdated() is still notified after the transition to
allow groups not to handle the migration. This changes iOS behavior
which will update its group on remote transition.

Bug: 370745855
Change-Id: Ica8389ff776454495548576ebaab26b81e5b3eae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6012818
Reviewed-by: Aliona Dangla <alionadangla@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Commit-Queue: Rushan Suleymanov <rushans@google.com>
Cr-Commit-Position: refs/heads/main@{#1384826}