Add a feature for form extraction improvements
This CL adds a wrapper class around base::optional_ref that ensures that
we only consider that an optional has a value if a feature is enabled.

Also the CL adds the feature in question and modifies the actual methods
that use this pattern.

Reason is we want to evaluate our metrics subject to this change and we
would need an experiment for this.

Low-Coverage-Reason: TRIVIAL_CHANGE this CL only adds a trivial wrapper around base::optional_ref.
Bug: 40947729
Change-Id: Iccbdbf3ca8260f220b5f3851795cc8736163f9c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6022836
Reviewed-by: Christoph Schwering <schwering@google.com>
Commit-Queue: Jihad Hanna <jihadghanna@google.com>
Cr-Commit-Position: refs/heads/main@{#1384861}