Remove redundant extraction operations while tracking submission
This CL moves the call to UpdateLastInteractedElement from
AutofillAgent::OnProvisionallySaveForm to the only place where it would
not be called and removes the calls from the function.

Previously, editing a field would induce two extraction operations: one
in FormTracker::FormControlDidChange and the other in
AutofillAgent::OnProvisionallySaveForm, even though the calls are back
to back.

Moreover, the CL uses the extracted form in `WillSendSubmitEvent` in
order to save an additional extraction operation in
InformBrowserAboutUserInput.

Bug: 40947729, 40281981
Change-Id: I4617c181826671fce594ffeed036aacb4352f476
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6001671
Commit-Queue: Jihad Hanna <jihadghanna@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#1384866}