Ignore extract options in the new extraction experiment
This CL makes all extraction operations assume that all extraction
options were passed to them.

The reason is that if in the future we want to use previous results of
operations to avoid calling extraction multiple times, the business of
having different extraction operations do multiple things would be
problematic.

The reason for having these operations in the first place was out of
worry about performance, since additional steps could require
additional time. However, looking at the metrics [1], we can see that
those two functions (one that extracts without any option [2] and the
other that extracts with all options [3]) don't really differ in time,
meaning that the difference in performance should be negligible.

Also note that in iOS for example, we always extract everything as of
recently [4]

[1] https://uma.googleplex.com/p/chrome/timeline_v2?sid=aecfaaad21e6b39e9885bae3d7b85171

[2] https://source.chromium.org/chromium/chromium/src/+/main:components/autofill/content/renderer/form_tracker.cc;l=458-462;drc=cfb0f1f104f7e7f1708b54322983a7c7d1321db4

[3] https://source.chromium.org/chromium/chromium/src/+/main:components/autofill/content/renderer/autofill_agent.cc;l=1338-1342;drc=60f3174c5de65c92f43aec2c513af7b7b4ed8de4

[4] https://chromium-review.googlesource.com/c/chromium/src/+/6025842?tab=comments#:~:text=removal%20might%20be-,parity%20with%20iOS,-.

Bug: 40947729
Change-Id: I4f629409fea02d066c35946b93f405f82880a6fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6025842
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Christoph Schwering <schwering@google.com>
Commit-Queue: Jihad Hanna <jihadghanna@google.com>
Cr-Commit-Position: refs/heads/main@{#1384876}